+++ modules/pytket-pyzx/_metadata.py
@@ -1,2 +1,2 @@
-__extension_version__ = "0.19.0"
+__extension_version__ = "0.20.0"
 __extension_name__ = "pytket-pyzx"