+++ modules/pytket-braket/docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 -------------------
 
 * Allow user to specify different local backends.
+* Fix dimension mismatch for measurements for Braket QPU results
 
 0.17.0 (March 2022)
 -------------------